requirements.txt
beautifulsoup4>=4.10.0
decorator>=5.1.0
future>=0.18.2
openpyxl>=3.0.9
psutil>=5.8.0
pydantic>=1.8.2
python_dateutil>=2.8.2
python_stdnum>=1.17
requests>=2.26.0
selenium>=4.1.0
setuptools>=59.3.0
tg.devtools>=2.4.3
tqdm>=4.62.3
